feat(CHK-2172): reset page after every test
