+++ e2e-tests/jest-puppeteer.config.js
@@ -1,9 +1,9 @@
 module.exports = {
-    launch: {
-        dumpio: true,
-        headless: "new",
-        product: 'chrome',
-        args: ["--no-sandbox"]
-    },
-    browserContext: 'incognito'
+  launch: {
+    dumpio: true,
+    headless: "new",
+    product: 'chrome',
+    args: ["--no-sandbox"]
+  },
+  browserContext: 'incognito'
 }
