+++ e2e-tests/package.json
@@ -15,7 +15,7 @@
     "@types/jest": "^29.5.7",
     "dotenv": "^16.3.1",
     "dotenv-cli": "^7.3.0",
-    "eslint": "^8.49.0",
+    "eslint": "^8.55.0",
     "jest": "^29.7.0",
     "jest-junit": "^16.0.0",
     "jest-puppeteer": "^9.0.0",
